[FLINK-3229] Move ShardConsumerThread out of KinesisDataFetcher as a public class for easier testing
